requirements.txt
--extra-index-url https://download.pytorch.org/whl/cu128

numpy>=1.23.5,<2
typing-extensions>=4.8.0
opencv-python==4.10.0.84
cv2_enumerate_cameras==1.1.15
onnx==1.16.0
insightface==0.7.3
psutil==5.9.8
tk==0.1.0
customtkinter==5.2.2
pillow==11.1.0
torch==2.10.0+cu128; sys_platform != 'darwin'
torch==2.10.0; sys_platform == 'darwin'
torchvision==0.25.0+cu128; sys_platform != 'darwin'
torchvision==0.25.0; sys_platform == 'darwin'
onnxruntime-silicon==1.19.2; sys_platform == 'darwin' and platform_machine == 'arm64'
onnxruntime-gpu==1.19.2; sys_platform != 'darwin'
tensorflow-cpu; sys_platform != 'darwin'
opennsfw2==0.10.2
protobuf==4.23.2
tqdm==4.66.4
# gfpgan package removed in favor of native ONNX implementation
tkinterdnd2==0.4.2
pygrabber==0.2
